Masked SlackMessageService behind an interface
